change transaction for data access object and service

+++ src/main/java/org/example/e_com/model/Product.java
@@ -1,13 +1,15 @@
 package org.example.e_com.model;
 
+import java.math.BigDecimal;
+
 public class Product {
     private int id;
     private String name;
-    private double price;
+    private BigDecimal price;
     private int quantity;
     private int category_id;
     public Product(){};
-    public Product(int id, String name, double price, int quantity, int category_id){
+    public Product(int id, String name, BigDecimal price, int quantity, int category_id){
         this.id = id;
         this.name = name;
         this.price = price;
@@ -31,11 +33,11 @@ public class Product {
         this.name = name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
